HW3/requirements.txt
functions-framework==3.*
Flask
google-cloud-storage
google-cloud-logging
google-cloud-pubsub
